JobMarketTracker/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
scrapy==2.11.0
celery==5.3.4
redis==5.0.1
pandas==2.1.3
spacy==3.7.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
django-filter==23.5
scrapy-djangoitem==1.1.1

